client/requirements.txt
certifi==2019.3.9
Click==7.0
cloudinary==1.15.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mock==3.0.5
six==1.12.0
urllib3==1.25.2
Werkzeug==0.15.2
grpcio==1.20.1
grpcio-tools==1.20.1
